requirements.txt
openai
python-dotenv
requests
pillow
pyngrok
moviepy
librosa
numpy
soundfile
ollama
pytest
